requirements.txt
click==6.6
Flask==0.11.1
geopy==1.11.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyMySQL==0.7.9
python-memcached==1.58
six==1.10.0
Werkzeug==0.11.11
wheel==0.24.0
